Started on Changes tasked by Mentor
