+++ src/main/java/lv/ctco/javaschool/auth/control/UserStore.java
@@ -78,13 +78,4 @@ public class UserStore {
             throw new InvalidPasswordException();
         }
     }
-
-    public UserLoginDto convertToDto(User user) {
-        UserLoginDto dto = new UserLoginDto();
-        dto.setUsername(user.getUsername());
-        dto.setPhone(user.getPhone());
-        dto.setEmail(user.getEmail());
-        return dto;
-    }
-
 }
