+++ src/test/java/lv/ctco/javaschool/auth/boundary/AuthenticationApiTest.java
@@ -3,6 +3,7 @@ package lv.ctco.javaschool.auth.boundary;
 import lv.ctco.javaschool.auth.control.UserStore;
 import lv.ctco.javaschool.auth.entity.domain.User;
 import lv.ctco.javaschool.auth.entity.dto.UserLoginDto;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -11,7 +12,6 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-
 
 class AuthenticationApiTest {
     @Mock
@@ -33,15 +33,23 @@ class AuthenticationApiTest {
         user.setEmail("a@b.com");
         Mockito.when(userStore.getCurrentUser()).thenReturn(user);
 
-        UserLoginDto dto = new UserLoginDto();
-        dto.setUsername(user.getUsername());
-        dto.setPhone(user.getPhone());
-        dto.setEmail(user.getEmail());
-        Mockito.when(userStore.convertToDto(user)).thenReturn(dto);
-
         UserLoginDto resultDto = authenticationApi.returnUserDto();
         assertEquals(user.getUsername(), resultDto.getUsername());
         assertEquals(user.getEmail(), resultDto.getEmail());
         assertEquals(user.getPhone(), resultDto.getPhone());
     }
+
+    @Test
+    void testUserConverterToDto() {
+        UserStore uc = new UserStore();
+        User user = new User();
+        user.setUsername("aa");
+        user.setPhone("1234567");
+        user.setEmail("a@b.com");
+        UserLoginDto dto = authenticationApi.convertToDto(user);
+        assertEquals(user.getUsername(), dto.getUsername());
+        assertEquals(user.getEmail(), dto.getEmail());
+        assertEquals(user.getPhone(), dto.getPhone());
+    }
+
 }
